client/detector/requirements.txt
numpy==1.16.3
paho-mqtt==1.4.0
